rename: rename page name into feed list page
